storageccl: delete ImportRequest as it is no longer used by restore
In 21.1 RESTORE stopped using the KV level ImportRequest for ingesting
data. The logic for restore was pulled into the data ingestion processor
as it did not rely on any KV specific information.

This change deletes all references of ImportRequest, and pulls out
some proto definitions that were previously nested in the ImportRequest.

Release note: None
