+++ pkg/ccl/backupccl/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
         "backup_processor.go",
         "backup_processor_planning.go",
         "create_scheduled_backup.go",
+        "key_rewriter.go",
         "manifest_handling.go",
         "restoration_data.go",
         "restore_data_processor.go",
@@ -128,8 +129,10 @@ go_test(
         "full_cluster_backup_restore_test.go",
         "helpers_test.go",
         "import_spans_test.go",
+        "key_rewriter_test.go",
         "main_test.go",
         "partitioned_backup_test.go",
+        "restore_data_processor_test.go",
         "restore_mid_schema_change_test.go",
         "restore_old_sequences_test.go",
         "restore_old_versions_test.go",
@@ -157,6 +160,7 @@ go_test(
         "//pkg/kv",
         "//pkg/kv/kvclient/kvcoord",
         "//pkg/kv/kvserver",
+        "//pkg/kv/kvserver/kvserverbase",
         "//pkg/roachpb",
         "//pkg/scheduledjobs",
         "//pkg/security",
@@ -193,6 +197,7 @@ go_test(
         "//pkg/testutils/testcluster",
         "//pkg/util",
         "//pkg/util/ctxgroup",
+        "//pkg/util/encoding",
         "//pkg/util/hlc",
         "//pkg/util/leaktest",
         "//pkg/util/log",
@@ -209,6 +214,7 @@ go_test(
         "@com_github_cockroachdb_datadriven//:datadriven",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_errors//oserror",
+        "@com_github_cockroachdb_pebble//vfs",
         "@com_github_gogo_protobuf//proto",
         "@com_github_gogo_protobuf//types",
         "@com_github_gorhill_cronexpr//:cronexpr",
