+++ pkg/ccl/storageccl/BUILD.bazel
@@ -5,8 +5,8 @@ go_library(
     srcs = [
         "encryption.go",
         "export.go",
+        "external_sst_reader.go",
         "import.go",
-        "key_rewriter.go",
         "revision_reader.go",
         "writebatch.go",
     ],
@@ -17,7 +17,6 @@ go_library(
         "//pkg/ccl/storageccl/engineccl",
         "//pkg/keys",
         "//pkg/kv",
-        "//pkg/kv/bulk",
         "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/batcheval",
         "//pkg/kv/kvserver/batcheval/result",
@@ -25,19 +24,14 @@ go_library(
         "//pkg/roachpb",
         "//pkg/settings",
         "//pkg/settings/cluster",
-        "//pkg/sql/catalog",
-        "//pkg/sql/catalog/descpb",
-        "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/sem/builtins",
         "//pkg/storage",
         "//pkg/storage/cloud",
         "//pkg/storage/enginepb",
-        "//pkg/util/encoding",
         "//pkg/util/hlc",
         "//pkg/util/humanizeutil",
         "//pkg/util/iterutil",
         "//pkg/util/log",
-        "//pkg/util/protoutil",
         "//pkg/util/retry",
         "//pkg/util/tracing",
         "@com_github_cockroachdb_errors//:errors",
@@ -53,8 +47,7 @@ go_test(
     srcs = [
         "encryption_test.go",
         "export_test.go",
-        "import_test.go",
-        "key_rewriter_test.go",
+        "external_sst_reader_test.go",
         "main_test.go",
         "writebatch_test.go",
     ],
@@ -63,33 +56,23 @@ go_test(
         "//pkg/base",
         "//pkg/keys",
         "//pkg/kv",
-        "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/batcheval",
-        "//pkg/kv/kvserver/kvserverbase",
         "//pkg/roachpb",
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/server",
         "//pkg/settings/cluster",
-        "//pkg/sql/catalog/descpb",
-        "//pkg/sql/catalog/systemschema",
-        "//pkg/sql/catalog/tabledesc",
-        "//pkg/sql/rowenc",
         "//pkg/storage",
-        "//pkg/storage/cloud",
         "//pkg/storage/cloudimpl",
         "//pkg/storage/enginepb",
         "//pkg/testutils",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
-        "//pkg/util/encoding",
         "//pkg/util/hlc",
         "//pkg/util/humanizeutil",
         "//pkg/util/leaktest",
-        "//pkg/util/protoutil",
         "//pkg/util/randutil",
-        "@com_github_cockroachdb_pebble//vfs",
         "@com_github_stretchr_testify//require",
     ],
 )
