+++ pkg/kv/batch.go
@@ -254,7 +254,6 @@ func (b *Batch) fillResults(ctx context.Context) {
 			case *roachpb.RequestLeaseRequest:
 			case *roachpb.CheckConsistencyRequest:
 			case *roachpb.WriteBatchRequest:
-			case *roachpb.ImportRequest:
 			case *roachpb.AdminScatterRequest:
 			case *roachpb.AddSSTableRequest:
 			case *roachpb.MigrateRequest:
