+++ pkg/kv/kvserver/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
         "addressing.go",
         "allocator.go",
         "allocator_scorer.go",
-        "cclglue.go",
         "compact_span_client.go",
         "consistency_queue.go",
         "debug_print.go",
