+++ pkg/roachpb/batch_generated.go
@@ -144,8 +144,6 @@ func (ru RequestUnion) GetInner() Request {
 		return t.WriteBatch
 	case *RequestUnion_Export:
 		return t.Export
-	case *RequestUnion_Import:
-		return t.Import
 	case *RequestUnion_QueryTxn:
 		return t.QueryTxn
 	case *RequestUnion_QueryIntent:
@@ -240,8 +238,6 @@ func (ru ResponseUnion) GetInner() Response {
 		return t.WriteBatch
 	case *ResponseUnion_Export:
 		return t.Export
-	case *ResponseUnion_Import:
-		return t.Import
 	case *ResponseUnion_QueryTxn:
 		return t.QueryTxn
 	case *ResponseUnion_QueryIntent:
@@ -409,8 +405,6 @@ func (ru *RequestUnion) MustSetInner(r Request) {
 		union = &RequestUnion_WriteBatch{t}
 	case *ExportRequest:
 		union = &RequestUnion_Export{t}
-	case *ImportRequest:
-		union = &RequestUnion_Import{t}
 	case *QueryTxnRequest:
 		union = &RequestUnion_QueryTxn{t}
 	case *QueryIntentRequest:
@@ -508,8 +502,6 @@ func (ru *ResponseUnion) MustSetInner(r Response) {
 		union = &ResponseUnion_WriteBatch{t}
 	case *ExportResponse:
 		union = &ResponseUnion_Export{t}
-	case *ImportResponse:
-		union = &ResponseUnion_Import{t}
 	case *QueryTxnResponse:
 		union = &ResponseUnion_QueryTxn{t}
 	case *QueryIntentResponse:
@@ -538,7 +530,7 @@ func (ru *ResponseUnion) MustSetInner(r Response) {
 	ru.Value = union
 }
 
-type reqCounts [45]int32
+type reqCounts [44]int32
 
 // getReqCounts returns the number of times each
 // request type appears in the batch.
@@ -612,30 +604,28 @@ func (ba *BatchRequest) getReqCounts() reqCounts {
 			counts[31]++
 		case *RequestUnion_Export:
 			counts[32]++
-		case *RequestUnion_Import:
-			counts[33]++
 		case *RequestUnion_QueryTxn:
-			counts[34]++
+			counts[33]++
 		case *RequestUnion_QueryIntent:
-			counts[35]++
+			counts[34]++
 		case *RequestUnion_AdminScatter:
-			counts[36]++
+			counts[35]++
 		case *RequestUnion_AddSstable:
-			counts[37]++
+			counts[36]++
 		case *RequestUnion_RecomputeStats:
-			counts[38]++
+			counts[37]++
 		case *RequestUnion_Refresh:
-			counts[39]++
+			counts[38]++
 		case *RequestUnion_RefreshRange:
-			counts[40]++
+			counts[39]++
 		case *RequestUnion_Subsume:
-			counts[41]++
+			counts[40]++
 		case *RequestUnion_RangeStats:
-			counts[42]++
+			counts[41]++
 		case *RequestUnion_AdminVerifyProtectedTimestamp:
-			counts[43]++
+			counts[42]++
 		case *RequestUnion_Migrate:
-			counts[44]++
+			counts[43]++
 		default:
 			panic(fmt.Sprintf("unsupported request: %+v", ru))
 		}
@@ -677,7 +667,6 @@ var requestNames = []string{
 	"LeaseInfo",
 	"WriteBatch",
 	"Export",
-	"Import",
 	"QueryTxn",
 	"QueryIntent",
 	"AdmScatter",
@@ -852,10 +841,6 @@ type exportResponseAlloc struct {
 	union ResponseUnion_Export
 	resp  ExportResponse
 }
-type importResponseAlloc struct {
-	union ResponseUnion_Import
-	resp  ImportResponse
-}
 type queryTxnResponseAlloc struct {
 	union ResponseUnion_QueryTxn
 	resp  QueryTxnResponse
@@ -943,18 +928,17 @@ func (ba *BatchRequest) CreateReply() *BatchResponse {
 	var buf30 []leaseInfoResponseAlloc
 	var buf31 []writeBatchResponseAlloc
 	var buf32 []exportResponseAlloc
-	var buf33 []importResponseAlloc
-	var buf34 []queryTxnResponseAlloc
-	var buf35 []queryIntentResponseAlloc
-	var buf36 []adminScatterResponseAlloc
-	var buf37 []addSSTableResponseAlloc
-	var buf38 []recomputeStatsResponseAlloc
-	var buf39 []refreshResponseAlloc
-	var buf40 []refreshRangeResponseAlloc
-	var buf41 []subsumeResponseAlloc
-	var buf42 []rangeStatsResponseAlloc
-	var buf43 []adminVerifyProtectedTimestampResponseAlloc
-	var buf44 []migrateResponseAlloc
+	var buf33 []queryTxnResponseAlloc
+	var buf34 []queryIntentResponseAlloc
+	var buf35 []adminScatterResponseAlloc
+	var buf36 []addSSTableResponseAlloc
+	var buf37 []recomputeStatsResponseAlloc
+	var buf38 []refreshResponseAlloc
+	var buf39 []refreshRangeResponseAlloc
+	var buf40 []subsumeResponseAlloc
+	var buf41 []rangeStatsResponseAlloc
+	var buf42 []adminVerifyProtectedTimestampResponseAlloc
+	var buf43 []migrateResponseAlloc
 
 	for i, r := range ba.Requests {
 		switch r.GetValue().(type) {
@@ -1189,90 +1173,83 @@ func (ba *BatchRequest) CreateReply() *BatchResponse {
 			buf32[0].union.Export = &buf32[0].resp
 			br.Responses[i].Value = &buf32[0].union
 			buf32 = buf32[1:]
-		case *RequestUnion_Import:
+		case *RequestUnion_QueryTxn:
 			if buf33 == nil {
-				buf33 = make([]importResponseAlloc, counts[33])
+				buf33 = make([]queryTxnResponseAlloc, counts[33])
 			}
-			buf33[0].union.Import = &buf33[0].resp
+			buf33[0].union.QueryTxn = &buf33[0].resp
 			br.Responses[i].Value = &buf33[0].union
 			buf33 = buf33[1:]
-		case *RequestUnion_QueryTxn:
+		case *RequestUnion_QueryIntent:
 			if buf34 == nil {
-				buf34 = make([]queryTxnResponseAlloc, counts[34])
+				buf34 = make([]queryIntentResponseAlloc, counts[34])
 			}
-			buf34[0].union.QueryTxn = &buf34[0].resp
+			buf34[0].union.QueryIntent = &buf34[0].resp
 			br.Responses[i].Value = &buf34[0].union
 			buf34 = buf34[1:]
-		case *RequestUnion_QueryIntent:
+		case *RequestUnion_AdminScatter:
 			if buf35 == nil {
-				buf35 = make([]queryIntentResponseAlloc, counts[35])
+				buf35 = make([]adminScatterResponseAlloc, counts[35])
 			}
-			buf35[0].union.QueryIntent = &buf35[0].resp
+			buf35[0].union.AdminScatter = &buf35[0].resp
 			br.Responses[i].Value = &buf35[0].union
 			buf35 = buf35[1:]
-		case *RequestUnion_AdminScatter:
+		case *RequestUnion_AddSstable:
 			if buf36 == nil {
-				buf36 = make([]adminScatterResponseAlloc, counts[36])
+				buf36 = make([]addSSTableResponseAlloc, counts[36])
 			}
-			buf36[0].union.AdminScatter = &buf36[0].resp
+			buf36[0].union.AddSstable = &buf36[0].resp
 			br.Responses[i].Value = &buf36[0].union
 			buf36 = buf36[1:]
-		case *RequestUnion_AddSstable:
+		case *RequestUnion_RecomputeStats:
 			if buf37 == nil {
-				buf37 = make([]addSSTableResponseAlloc, counts[37])
+				buf37 = make([]recomputeStatsResponseAlloc, counts[37])
 			}
-			buf37[0].union.AddSstable = &buf37[0].resp
+			buf37[0].union.RecomputeStats = &buf37[0].resp
 			br.Responses[i].Value = &buf37[0].union
 			buf37 = buf37[1:]
-		case *RequestUnion_RecomputeStats:
+		case *RequestUnion_Refresh:
 			if buf38 == nil {
-				buf38 = make([]recomputeStatsResponseAlloc, counts[38])
+				buf38 = make([]refreshResponseAlloc, counts[38])
 			}
-			buf38[0].union.RecomputeStats = &buf38[0].resp
+			buf38[0].union.Refresh = &buf38[0].resp
 			br.Responses[i].Value = &buf38[0].union
 			buf38 = buf38[1:]
-		case *RequestUnion_Refresh:
+		case *RequestUnion_RefreshRange:
 			if buf39 == nil {
-				buf39 = make([]refreshResponseAlloc, counts[39])
+				buf39 = make([]refreshRangeResponseAlloc, counts[39])
 			}
-			buf39[0].union.Refresh = &buf39[0].resp
+			buf39[0].union.RefreshRange = &buf39[0].resp
 			br.Responses[i].Value = &buf39[0].union
 			buf39 = buf39[1:]
-		case *RequestUnion_RefreshRange:
+		case *RequestUnion_Subsume:
 			if buf40 == nil {
-				buf40 = make([]refreshRangeResponseAlloc, counts[40])
+				buf40 = make([]subsumeResponseAlloc, counts[40])
 			}
-			buf40[0].union.RefreshRange = &buf40[0].resp
+			buf40[0].union.Subsume = &buf40[0].resp
 			br.Responses[i].Value = &buf40[0].union
 			buf40 = buf40[1:]
-		case *RequestUnion_Subsume:
+		case *RequestUnion_RangeStats:
 			if buf41 == nil {
-				buf41 = make([]subsumeResponseAlloc, counts[41])
+				buf41 = make([]rangeStatsResponseAlloc, counts[41])
 			}
-			buf41[0].union.Subsume = &buf41[0].resp
+			buf41[0].union.RangeStats = &buf41[0].resp
 			br.Responses[i].Value = &buf41[0].union
 			buf41 = buf41[1:]
-		case *RequestUnion_RangeStats:
+		case *RequestUnion_AdminVerifyProtectedTimestamp:
 			if buf42 == nil {
-				buf42 = make([]rangeStatsResponseAlloc, counts[42])
+				buf42 = make([]adminVerifyProtectedTimestampResponseAlloc, counts[42])
 			}
-			buf42[0].union.RangeStats = &buf42[0].resp
+			buf42[0].union.AdminVerifyProtectedTimestamp = &buf42[0].resp
 			br.Responses[i].Value = &buf42[0].union
 			buf42 = buf42[1:]
-		case *RequestUnion_AdminVerifyProtectedTimestamp:
+		case *RequestUnion_Migrate:
 			if buf43 == nil {
-				buf43 = make([]adminVerifyProtectedTimestampResponseAlloc, counts[43])
+				buf43 = make([]migrateResponseAlloc, counts[43])
 			}
-			buf43[0].union.AdminVerifyProtectedTimestamp = &buf43[0].resp
+			buf43[0].union.Migrate = &buf43[0].resp
 			br.Responses[i].Value = &buf43[0].union
 			buf43 = buf43[1:]
-		case *RequestUnion_Migrate:
-			if buf44 == nil {
-				buf44 = make([]migrateResponseAlloc, counts[44])
-			}
-			buf44[0].union.Migrate = &buf44[0].resp
-			br.Responses[i].Value = &buf44[0].union
-			buf44 = buf44[1:]
 		default:
 			panic(fmt.Sprintf("unsupported request: %+v", r))
 		}
@@ -1349,8 +1326,6 @@ func CreateRequest(method Method) Request {
 		return &WriteBatchRequest{}
 	case Export:
 		return &ExportRequest{}
-	case Import:
-		return &ImportRequest{}
 	case QueryTxn:
 		return &QueryTxnRequest{}
 	case QueryIntent:
