+++ pkg/sql/execinfrapb/processors_bulk_io.proto
@@ -188,20 +188,34 @@ message BackupDataSpec {
   optional string user_proto = 10 [(gogoproto.nullable) = false, (gogoproto.casttype) = "github.com/cockroachdb/cockroach/pkg/security.SQLUsernameProto"];
 }
 
+message RestoreFileSpec {
+  optional roachpb.ExternalStorage dir = 1 [(gogoproto.nullable) = false];
+  optional string path = 2 [(gogoproto.nullable) = false];
+  reserved 3;
+  optional bytes sha512 = 4;
+}
+
+message TableRekey {
+  // OldID is the previous ID of `new_desc`.
+  optional uint32 old_id = 1 [(gogoproto.nullable) = false, (gogoproto.customname) = "OldID"];
+  // NewDesc is an encoded Descriptor message.
+  optional bytes new_desc = 2;
+}
+
 // RestoreDataEntry will be specified at planning time to the SplitAndScatter
 // processors, then those processors will stream these, encoded as bytes in rows
 // to the RestoreDataProcessors.
 // This field has a subset of the importEntry struct defined in restore.
 message RestoreSpanEntry {
   optional roachpb.Span span = 1 [(gogoproto.nullable) = false];
-  repeated roachpb.ImportRequest.File files = 2 [(gogoproto.nullable) = false];
+  repeated RestoreFileSpec files = 2 [(gogoproto.nullable) = false];
   optional int64 progressIdx = 3 [(gogoproto.nullable) = false];
 }
 
 message RestoreDataSpec {
   optional util.hlc.Timestamp restore_time = 1 [(gogoproto.nullable) = false];
   optional roachpb.FileEncryptionOptions encryption = 2;
-  repeated roachpb.ImportRequest.TableRekey rekeys = 3 [(gogoproto.nullable) = false];
+  repeated TableRekey rekeys = 3 [(gogoproto.nullable) = false];
 
   // PKIDs is used to convert result from an ExportRequest into row count
   // information passed back to track progress in the backup job.
@@ -214,7 +228,7 @@ message SplitAndScatterSpec {
   }
 
   repeated RestoreEntryChunk chunks = 1 [(gogoproto.nullable) = false];
-  repeated roachpb.ImportRequest.TableRekey rekeys = 2 [(gogoproto.nullable) = false];
+  repeated TableRekey rekeys = 2 [(gogoproto.nullable) = false];
 }
 
 // FileCompression list of the compression codecs which are currently
